backend/requirements.txt
fastapi
uvicorn
feedparser
beautifulsoup4
python-dotenv
google-generativeai
pymongo
requests
PyJWT
passlib
bcrypt